Add e2e for Android and move common code between Android and JVM to a different package
