+++ iot-e2e-tests/android/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Mar 09 15:59:14 PST 2017
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-3.3-all.zip
