+++ iot-e2e-tests/android/settings.gradle
@@ -1,0 +1,1 @@
+include ':app'
