+++ iot-e2e-tests/common/src/main/java/com/microsoft/azure/sdk/iot/common/Success.java
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+package com.microsoft.azure.sdk.iot.common;
+
+public class Success
+{
+    public Boolean result = false;
+
+    public void setResult(Boolean result)
+    {
+        this.result = result;
+    }
+
+    public Boolean getResult()
+    {
+        return this.result;
+    }
+}
